program files
The files I'm adding are programs, the first one is just a hard coded program for 3 digit numbers but after that I made a second version that can get the narcissistic numbers by brute-forcing. I will use this program to get a big sample size to study the xtics of these numbers and formulate a formula to get these numbers directly. I have seen that the version 2 program only handles numbers up to 100,000,000. Be careful because when I used this the computer almost crashed. I only have 8 GB of ram so this might be that case but I will try using Numpy arrays instead of lists to see if the memory issue will be resolved . If not I will try using batch processing and other methods to solve the memory issue to enable the program to work with big numbers like like a googleplex. I will upload my notes on my findings later.