final regular program uploads.
The files that I am uploading pertain to the program for generating narcissistic numbers. This was just part of the project. I want to make a formula for producing narcissistic numbers. I will be posting the Machine Learning part of this since this is the way that I decided to find the formula. I will also be working on some pure observation methods but I am counting on the Machine Learning aspect to give me more finite results. It will also be a very good workout on Machine Learning/ Data Science. About the results:
The laptop is a 2 core 4 thread CPU and has 8 GB of RAM. the python and compiler specifics are on the screenshots of the results.