Output and missing information.
This is the output of the two programs. The one for Tuesday is the Idle output and the one for Wednesday is the Spyder 5 output. the missing info is that i got the batch processing to work. It was the only way that the program could work on my laptop since when I tried to run it straight from the full array to the loop part of my program, it froze my computer and I had to restart the entire machine. I will finally try to use it to run an array with 1,000,000,000,000 (Trillion). I will be posting the results later.